refactor: rename perm_insure.views
